FIX: winter tile change for whole map also on large maps
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1463 8aca7d54-2c30-db11-9de9-000461428c89

+++ simworld.cc
@@ -2131,7 +2131,7 @@ karte_t::step()
 				INT_CHECK("karte_t::step");
 			}
 		}
-		if(  tile_counter >= (uint16)(cached_groesse_gitter_x*cached_groesse_gitter_y)  ) {
+		if(  tile_counter >= (uint32)cached_groesse_gitter_x*(uint32)cached_groesse_gitter_y  ) {
 			pending_season_change --;
 			tile_counter = 0;
 		}
